fix: update to new delete-confirmation modal
