+++ src/graphql/index.js
@@ -1,5 +1,6 @@
 export * from "./create-user";
 export * from "./delete-fest";
+export * from "./delete-user";
 export * from "./get-fest";
 export * from "./get-movies";
 export * from "./get-user-authentication";
